refactored metrics tests into one file
